code_800F7260 (#433)
* Split code_800F7260

* More function splits, migrate rodata

* Initial progress

* Small progress

* More substantial progress, finally figured out structs

* 2 functions left

* Move prototypes to functions.h

* Remove unused asm, move most externs to variables.h

* Fix merge

* Ran formatter

* Brackets to avoid && &


* More variables to variables.h

* Braces around early return


* ARRAY_COUNT in sBankSizes and remove unnecessary temp vars

+++ src/overlays/actors/ovl_player_actor/z_player.c
@@ -1341,7 +1341,7 @@ void func_808326F0(Player* this) {
     s32 i;
 
     for (i = 0; i < 4; i++) {
-        func_800F8D04(*entry + this->ageProperties->unk_92);
+        func_800F8D04((u16)(*entry + this->ageProperties->unk_92));
         entry++;
     }
 }
